model/requirements.txt
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-mistralai>=0.2.0
langchain-core==1.2.5
python-dotenv==1.0.0
fastapi==0.121.3
requests==2.32.5
pyyaml==6.0.3
yandex-cloud-ml-sdk==0.17.1
langchain-community==0.4.1
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.24.1
tavily-python==0.7.18
httpx>=0.27.0
